% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allvariables.manual.review.R
\name{allvariables.of.type}
\alias{allvariables.of.type}
\title{Extracts all variables of a given from a pair list}
\usage{
allvariables.of.type(list, type)
}
\arguments{
\item{list}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{type}{The name of a type, e.g. \code{"numeric"}}
}
\value{
A vector of variable names
}
\description{
Extracts all variables of a given from a pair list
}
\examples{
list <- lapply(names(mtcars), function(v) c(v,class(mtcars[[v]])))
allvariables.of.type(list, "numeric")
}

