% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.manual.review.R
\name{allvariables.type.change}
\alias{allvariables.type.change}
\title{Changing a data frame's variable types in bulk}
\usage{
allvariables.type.change(data, vars.list, data2)
}
\arguments{
\item{data}{A data frame}

\item{vars.list}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{data2}{A second data frame (optional) from which levels will be pulled and added when coercing a variable into a factor}
}
\value{
A modified version of \code{data}
}
\description{
Changing a data frame's variable types in bulk
}
\examples{
list_ <- list(c("cyl","factor"),c("am","factor"),c("vs","factor"),c("gear","factor"))
str(allvariables.type.change(mtcars,list_))
}

