% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.manual.review.R
\name{change.variable.type.in.list}
\alias{change.variable.type.in.list}
\title{Manually changes an element in a (name,type) list}
\usage{
change.variable.type.in.list(list, varname, vartype)
}
\arguments{
\item{list}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{varname}{A variable name}

\item{vartype}{A variable type}
}
\value{
A modified version of \code{list}
}
\description{
Manually changes an element in a (name,type) list
}
\examples{
list <- lapply(names(mtcars), function(v) c(v,class(mtcars[[v]])) )
change.variable.type.in.list(list, "am", "factor")
}

