% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.combinations.R
\name{factor.combinations.test}
\alias{factor.combinations.test}
\title{Combines two factor variables and measures the new variable's relative info}
\usage{
factor.combinations.test(data, pair, var.output)
}
\arguments{
\item{data}{A data frame}

\item{pair}{A pair of variable names}

\item{var.output}{A variable name}
}
\value{
A data frame with each input variable and the combined variable's number of unique values and relative information
}
\description{
Combines two factor variables and measures the new variable's relative info
}

