% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.factor.info.R
\name{factor.factor.info}
\alias{factor.factor.info}
\title{Measuring relative information between factors and output variable}
\usage{
factor.factor.info(data, var.output, list)
}
\arguments{
\item{data}{A data frame}

\item{var.output}{The name of the variable output}

\item{list}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}
}
\value{
A data frame with two columns: variable names, and their relative information to the output variable
}
\description{
Measuring relative information between factors and output variable
}
\examples{
X <- data.frame(a = factor(sample(1:3,100,TRUE)),
                b = factor(sample(1:5,100,TRUE)),
                c = factor(sample(letters[1:5],100,TRUE)))
pairslist <- list(c("a","factor"), c("b","integer"), c("c","factor"))
factor.factor.info(X, "a", pairslist)
}

