% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.info.to.output.R
\name{factor.penal.info.to.output}
\alias{factor.penal.info.to.output}
\title{Measuring relative information between factors and output variable}
\usage{
factor.penal.info.to.output(data, var.output, vars.list, alpha = 2,
  return.list = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{var.output}{The name of the variable output}

\item{vars.list}{(\emph{optional}) A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{alpha}{(\emph{defaults to 2}) The penalty}

\item{return.list}{(\emph{defaults to TRUE}) If \code{TRUE}, the function returns a list of one-line dataframes; if \code{FALSE}, it returns a concatenated dataframe sorted by decreasing value of \code{RI.to.output}}
}
\value{
A data frame with three columns: variable names, the number of active levels, and their penalized relative information to the output variable
}
\description{
Measuring relative information between factors and output variable
}
\examples{
X <- data.frame(a = factor(sample(1:3,100,TRUE)),
                b = factor(sample(1:5,100,TRUE)),
                c = factor(sample(letters[1:5],100,TRUE)))
factor.penal.info.to.output(X, "a", alpha = 2)
list_ <- list(c("vs","factor"),c("am","factor"),c("gear","factor"))
factor.penal.info.to.output(mtcars, "cyl", vars.list = list_, alpha = 1.5, return.list = TRUE)
}

