% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.replace.NA.R
\name{factor.replace.NA}
\alias{factor.replace.NA}
\title{Replace NA in factor variables}
\usage{
factor.replace.NA(data, tag)
}
\arguments{
\item{data}{A data frame}

\item{tag}{A string}
}
\value{
A modified version of \code{data}
}
\description{
Replace NA in factor variables
}
\examples{
X <- data.frame(a = c(1,2,NA), b = c("a",NA,"c"))
factor.replace.NA(X,"unk")
}

