% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.skinning.R
\name{factor.skinning}
\alias{factor.skinning}
\title{Removes thin levels from factor variables in a data frame}
\usage{
factor.skinning(data, list, thresh = 0.001, tag = NA)
}
\arguments{
\item{data}{A data frame}

\item{list}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{thresh}{Defines the minimum number (or, if <1, the minimum proportion) of occurences of a level to be kept}

\item{tag}{A value that will replace levels that do not appear in both \code{data1[[variable]]} and \code{data2[[variable]]}}
}
\value{
A modified version of data
}
\description{
Removes thin levels from factor variables in a data frame
}
\examples{
set.seed(1)
X <- data.frame(a = factor(sample(1:5,100,TRUE)),
                b = factor(sample(letters[1:5],100,TRUE)))
pairslist <- list(c("a","factor"), c("b","factor"))
Y <- factor.skinning(X, pairslist, thresh = 0.19)
table(Y$a)
table(Y$b)
}

