% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.info.R
\name{penalized.relative.info}
\alias{penalized.relative.info}
\title{Penalized relative information}
\usage{
penalized.relative.info(data, var.x, var.y, alpha = 2)
}
\arguments{
\item{data}{A data frame}

\item{var.x}{The \emph{regressor} (or \emph{predictor}) variable (has to be the name of a variable in \code{data})}

\item{var.y}{The \emph{output} variable (has to be the name of a variable in \code{data})}

\item{alpha}{(\emph{defaults to 2}) A numeric}
}
\value{
PenRI(var.y|var.x;alpha) = Sum_(a ~ var.x) P(a)^(alpha) * PointRI(var.y|var.x = a)
}
\description{
Penalized relative information
}
\examples{
list.pairs <- list(c("cyl","factor"),c("am","factor"),c("vs","factor"),c("gear","factor"))
mtcars2 <- allvariables.type.change(mtcars, list.pairs)
penalized.relative.info(mtcars2,"cyl","gear",3/2)
}

