% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.info.R
\name{pointwise.relative.info}
\alias{pointwise.relative.info}
\title{Pointwise relative information}
\usage{
pointwise.relative.info(data, var.x, var.y, x.value)
}
\arguments{
\item{data}{A data frame with at least two variables}

\item{var.x}{The name of a variable in data}

\item{var.y}{The name of a variable in data}

\item{x.value}{The value of var.x to condition upon}
}
\value{
PRI(var.y | var.x = x.value) = 1 - H(var.y | var.x = x.value) / H(var.y)
}
\description{
Pointwise relative information
}
\examples{
pointwise.relative.info(data = mtcars, var.x = "cyl", var.y = "mpg", x.value = 4)
}

