% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.info.R
\name{relative.info}
\alias{relative.info}
\title{Relative information}
\usage{
relative.info(data, var.x, var.y, x.value)
}
\arguments{
\item{data}{A data frame}

\item{var.x}{The \emph{regressor} (or \emph{predictor}) variable (has to be the name of a variable in \code{data})}

\item{var.y}{The \emph{output} variable (has to be the name of a variable in \code{data})}

\item{x.value}{(\strong{optional}) The value of \code{var.x} to condition upon (has to be a value of \code{data[[var.x]]})}
}
\value{
\itemize{
\item If \code{x.value} \strong{is} specified, returns
PRI(var.y|var.x = x.value) = 1 - H(var.y|var.x = x.value)/H(var.y).
\item If \code{x.value} \strong{is not} specified, returns RI(var.y|var.x) = MI(var.x,var.y)/H(var.y)
}
}
\description{
Relative information
}
\examples{
relative.info(data = mtcars, var.x = "cyl", var.y = "mpg")
relative.info(data = mtcars, var.x = "cyl", var.y = "mpg", x.value = 4)
}

