% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.combinations.R
\name{select.combinations}
\alias{select.combinations}
\title{A helper function to filter a list of combinations}
\usage{
select.combinations(list, max.factors, min.RI)
}
\arguments{
\item{list}{A list such as those created by \code{factor.allpairs.test}}

\item{max.factors}{An integer}

\item{min.RI}{An integer}
}
\value{
A modified version of \code{list}
}
\description{
A helper function to filter a list of combinations
}
\examples{
list.pairs <- list(c("cyl","factor"),c("vs","factor"),c("am","factor"),
                   c("gear","factor"),c("carb","factor"))
list.combs  <- factor.allpairs.test(mtcars,pairs,"mpg")
list.combs2 <- select.combinations(list.combs, max.factors = 6, min.RI = 0.4)
}

