% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.info.R
\name{total.relative.info}
\alias{total.relative.info}
\title{Total relative information}
\usage{
total.relative.info(data, var.x, var.y)
}
\arguments{
\item{data}{A data frame with at least two variables}

\item{var.x}{The name of a variable in data}

\item{var.y}{The name of a variable in data}
}
\value{
RI(var.y | var.x) = MI(var.x,var.y)/H(var.y)
}
\description{
Total relative information
}
\examples{
total.relative.info(data = iris, var.x = "Sepal.Width", var.y = "Species")
total.relative.info(data = mtcars, var.x = "cyl", var.y = "mpg")
}

