% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.combinations.R
\name{view.allpairs.test.result}
\alias{view.allpairs.test.result}
\title{A helper function to view the output of factor.allpairs.test in a table format}
\usage{
view.allpairs.test.result(list)
}
\arguments{
\item{list}{A list such as those created by \code{factor.allpairs.test}}
}
\value{
A data frame sorted by decreasing values of \code{RI.combined}
}
\description{
A helper function to view the output of factor.allpairs.test in a table format
}
\examples{
pairs <- list(c("cyl","factor"),c("vs","factor"),c("am","factor"),
              c("gear","factor"),c("carb","factor"))
list  <- factor.allpairs.test(mtcars,pairs,"mpg")
View(view.allpairs.test.result(list))
}

