% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.variables.summary2.R
\name{allvariables.manual.review}
\alias{allvariables.manual.review}
\title{Manual review of a data frame's variables}
\usage{
allvariables.manual.review(data, clear.console = TRUE,
  max.unique.factor = 50)
}
\arguments{
\item{data}{A data frame}

\item{clear.console}{(\emph{defaults to TRUE}) If TRUE, the console is cleared before each new variable's summary is printed}

\item{max.unique.factor}{(\emph{defaults to 50}) When a variable's number of unique values is lower than this, the guessed type is \code{factor}}
}
\value{
A list of pairs (variable,type)
}
\description{
Manual review of a data frame's variables
}
\examples{
allvariables.manual.review(iris)   # all default guesses are good
allvariables.manual.review(mtcars) # some variables should be re-classified as integer or factor
}

