% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.type.change.R
\name{allvariables.type.change}
\alias{allvariables.type.change}
\title{Changing a data frame's variable types in bulk}
\usage{
allvariables.type.change(data, list, data2)
}
\arguments{
\item{data}{A data frame}

\item{list}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{data2}{A second data frame (optional) from which values will be pulled when coercing a variable into a factor}
}
\value{
A modified version of \code{data}
}
\description{
Changing a data frame's variable types in bulk
}

