% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.combinations.R
\name{factor.add.combinations}
\alias{factor.add.combinations}
\title{Adds combinations of factors to a data frame}
\usage{
factor.add.combinations(data, list.pairs, list.combs)
}
\arguments{
\item{data}{A data frame}

\item{list.pairs}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{list.combs}{A list such as those created by \code{factor.allpairs.test}}
}
\value{
If \code{list.pairs} is supplied, a list whose first element is an expanded version of \code{data} and whose second element is the corresponding expanded version of \code{list.pairs}; if \code{list.pairs} is missing, returns only the modified version of \code{data}
}
\description{
Adds combinations of factors to a data frame
}
\examples{
list.pairs <- list(c("cyl","factor"),c("vs","factor"),c("am","factor"),
                   c("gear","factor"),c("carb","factor"))
list.combs  <- factor.allpairs.test(mtcars,pairs,"mpg")
view.allpairs.test.result(list.combs)
newdata <- factor.add.combinations(data = mtcars, list.combs = list.combs)
head(newdata)
list.combs2 <- select.combinations(list.combs, max.factors = 6, min.RI = 0.4)
newdata2 <- factor.add.combinations(data = mtcars, list.pairs = list.pairs, list.combs = list.combs2)
head(newdata2[[1]])
tail(newdata2[[2]])
}

