% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.combinations.R
\name{factor.allpairs.test}
\alias{factor.allpairs.test}
\title{Creates all combinations of 2 factor variables within a given scope and measures their relative information}
\usage{
factor.allpairs.test(data, list, var.output, parallel = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{list}{A list of pairs (variable.name,variable.type) such as those produced by \code{allvariables.manual.review}}

\item{var.output}{The name of the variable output}

\item{parallel}{(\emph{defaults to FALSE}) If TRUE, dispatches factor pairs in parallel}
}
\value{
A list
}
\description{
Creates all combinations of 2 factor variables within a given scope and measures their relative information
}
\examples{
list.pairs <- list(c("cyl","factor"),c("vs","factor"),c("am","factor"),
                   c("gear","factor"),c("carb","factor"))
list.combs  <- factor.allpairs.test(mtcars,pairs,"mpg")
View(view.allpairs.test.result(list.combs))
}

